us-net-tc schemas moved to draft repository
The application schema Telecommunications network is only proposed in the Technical guidance and it is not included in the Implementing Rules.
To be in line with the scope of the "official" schema repository - contains only those data models that are contained in the amendment to the Implementing Rules for Annex II+III themes, including the updates of the Annex I data themes - the application schema Telecommunications network should not be present in the schema repository but only in the draft one.